requirements.txt
boto3
google-generativeai